Display a word's lexicalCategory alongwith the word name
lexicalCategory here means what part of speech the word is. It can be a
Noun, Verb, Adverb, Adjective or something else. For words with multiple
lexical categories, we randomly chose one lexical category and display
the meaning accordingly. For example, the word 'bad' is an adverb as
well as an adjective. Depending on the random number generated we will
either show the adverb entry or the adjective one.